chore(e2e): move extension-load code to helpers #556
